+++ src/wipeit.py
@@ -393,7 +393,7 @@ Examples:
         '-v',
         '--version',
         action='version',
-        version='wipeit 1.3.0')
+        version='wipeit 1.3.1')
 
     args = parser.parse_args()